+++ Editor/DLLs/Descriptors/GeometryImportSettings.cs
@@ -1,4 +1,5 @@
 ﻿using Editor.Content;
+using System.IO;
 using System.Runtime.InteropServices;
 
 namespace Editor.DLLs.Descriptors